get rid of submodules
