+++ c_fms/Makefile.am
@@ -27,20 +27,16 @@ AM_CPPFLAGS = -I. -I./include
 AM_FCFLAGS = $(FC_MODINC). $(FC_MODOUT)$(MODDIR)
 
 # Build these uninstalled convenience libraries.
-noinst_LTLIBRARIES = libcfms.la
+noinst_LTLIBRARIES = lib_c_fms.la
 
 # Each convenience library depends on its source.
-libcfms_la_SOURCES = cfms.F90 cmpp.F90 cmpp_domains.F90
-
-cfms_mod@cmpp_domains_smod.smod: cmpp_domains.F90 cfms_mod.mod cfms_mod.smod
-cfms_mod@cmpp_smod.smod: cmpp.F90 cfms_mod.mod cfms_mod.smod
-cfms_mod.mod, cfms_mod.smod: cfms.F90 
+lib_c_fms_la_SOURCES = c_fms.F90 
 
 # Mod files are built and then installed as headers.
-MODFILES = cfms_mod.mod #cfms_mod.smod cfms_mod@cmpp_smod.smod cfms_mod@cmpp_domains_smod.smod
+MODFILES = c_fms_mod.$(FC_MODEXT)
 BUILT_SOURCES = $(MODFILES)
 nodist_include_HEADERS =  $(FMS_INC_FILES) $(MODFILES)
 
-include_HEADERS = cfms.h cmpp.h cmpp_domains.h
+include_HEADERS = c_fms.h
 
 include $(top_srcdir)/mkmods.mk
