--- cfms/include/cfms_interfaces.fh
@@ -1,111 +1,0 @@
-! int 2d
-#undef  CFMS_UPDATE_DOMAINS_SUB_NAME_
-#undef  CFMS_UPDATE_DOMAINS_BINDC_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_NDIM_
-#define CFMS_UPDATE_DOMAINS_SUB_NAME_     cFMS_update_domains_int_2d
-#define CFMS_UPDATE_DOMAINS_BINDC_        "cFMS_update_domains_int_2d"
-#define CFMS_UPDATE_DOMAINS_FIELD_NDIM_   2
-#include "cfms_interfaces.inc"
-
-!int 3d
-#undef  CFMS_UPDATE_DOMAINS_SUB_NAME_
-#undef  CFMS_UPDATE_DOMAINS_BINDC_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_NDIM_
-#define CFMS_UPDATE_DOMAINS_SUB_NAME_     cFMS_update_domains_int_3d
-#define CFMS_UPDATE_DOMAINS_BINDC_        "cFMS_update_domains_int_3d"
-#define CFMS_UPDATE_DOMAINS_FIELD_NDIM_   3
-#include "cfms_interfaces.inc"
-  
-!int 4d
-#undef  CFMS_UPDATE_DOMAINS_SUB_NAME_
-#undef  CFMS_UPDATE_DOMAINS_BINDC_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_NDIM_
-#define CFMS_UPDATE_DOMAINS_SUB_NAME_     cFMS_update_domains_int_4d
-#define CFMS_UPDATE_DOMAINS_BINDC_        "cFMS_update_domains_int_4d"
-#define CFMS_UPDATE_DOMAINS_FIELD_NDIM_   4
-#include "cfms_interfaces.inc"
-  
-!int 5d
-#undef  CFMS_UPDATE_DOMAINS_SUB_NAME_
-#undef  CFMS_UPDATE_DOMAINS_BINDC_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_NDIM_
-#define CFMS_UPDATE_DOMAINS_SUB_NAME_     cFMS_update_domains_int_5d
-#define CFMS_UPDATE_DOMAINS_BINDC_        "cFMS_update_domains_int_5d"
-#define CFMS_UPDATE_DOMAINS_FIELD_NDIM_   5
-#include "cfms_interfaces.inc"
-
-!float 2d
-#undef  CFMS_UPDATE_DOMAINS_SUB_NAME_
-#undef  CFMS_UPDATE_DOMAINS_BINDC_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_NDIM_
-#define CFMS_UPDATE_DOMAINS_SUB_NAME_     cFMS_update_domains_float_2d
-#define CFMS_UPDATE_DOMAINS_BINDC_        "cFMS_update_domains_float_2d"
-#define CFMS_UPDATE_DOMAINS_FIELD_NDIM_   2
-#include "cfms_interfaces.inc"
-
-!float 3d
-#undef  CFMS_UPDATE_DOMAINS_SUB_NAME_
-#undef  CFMS_UPDATE_DOMAINS_BINDC_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_NDIM_
-#define CFMS_UPDATE_DOMAINS_SUB_NAME_     cFMS_update_domains_float_3d
-#define CFMS_UPDATE_DOMAINS_BINDC_        "cFMS_update_domains_float_3d"
-#define CFMS_UPDATE_DOMAINS_FIELD_NDIM_   3
-#include "cfms_interfaces.inc"
-  
-!float 4d
-#undef  CFMS_UPDATE_DOMAINS_SUB_NAME_
-#undef  CFMS_UPDATE_DOMAINS_BINDC_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_NDIM_
-#define CFMS_UPDATE_DOMAINS_SUB_NAME_     cFMS_update_domains_float_4d
-#define CFMS_UPDATE_DOMAINS_BINDC_        "cFMS_update_domains_float_4d"
-#define CFMS_UPDATE_DOMAINS_FIELD_NDIM_   4
-#include "cfms_interfaces.inc"
-  
-!float 5d
-#undef  CFMS_UPDATE_DOMAINS_SUB_NAME_
-#undef  CFMS_UPDATE_DOMAINS_BINDC_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_NDIM_
-#define CFMS_UPDATE_DOMAINS_SUB_NAME_     cFMS_update_domains_float_5d
-#define CFMS_UPDATE_DOMAINS_BINDC_        "cFMS_update_domains_float_5d"
-#define CFMS_UPDATE_DOMAINS_FIELD_NDIM_   5
-#include "cfms_interfaces.inc"
-  
-!double 2d
-#undef  CFMS_UPDATE_DOMAINS_SUB_NAME_
-#undef  CFMS_UPDATE_DOMAINS_BINDC_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_NDIM_
-#define CFMS_UPDATE_DOMAINS_SUB_NAME_     cFMS_update_domains_double_2d
-#define CFMS_UPDATE_DOMAINS_BINDC_        "cFMS_update_domains_double_2d"
-#define CFMS_UPDATE_DOMAINS_FIELD_NDIM_   2
-#include "cfms_interfaces.inc"
-
-!double 3d
-#undef  CFMS_UPDATE_DOMAINS_SUB_NAME_
-#undef  CFMS_UPDATE_DOMAINS_BINDC_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_NDIM_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_POINTER_F
-#define CFMS_UPDATE_DOMAINS_SUB_NAME_     cFMS_update_domains_double_3d
-#define CFMS_UPDATE_DOMAINS_BINDC_        "cFMS_update_domains_double_3d"
-#define CFMS_UPDATE_DOMAINS_FIELD_NDIM_   3
-#include "cfms_interfaces.inc"
-  
-!double 4d
-#undef  CFMS_UPDATE_DOMAINS_SUB_NAME_
-#undef  CFMS_UPDATE_DOMAINS_BINDC_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_NDIM_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_POINTER_F
-#define CFMS_UPDATE_DOMAINS_SUB_NAME_     cFMS_update_domains_double_4d
-#define CFMS_UPDATE_DOMAINS_BINDC_        "cFMS_update_domains_double_4d"
-#define CFMS_UPDATE_DOMAINS_FIELD_NDIM_   4
-#include "cfms_interfaces.inc"
-  
-!double 5d
-#undef  CFMS_UPDATE_DOMAINS_SUB_NAME_
-#undef  CFMS_UPDATE_DOMAINS_BINDC_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_NDIM_
-#undef  CFMS_UPDATE_DOMAINS_FIELD_POINTER_F
-#define CFMS_UPDATE_DOMAINS_SUB_NAME_     cFMS_update_domains_double_5d
-#define CFMS_UPDATE_DOMAINS_BINDC_        "cFMS_update_domains_double_5d"
-#define CFMS_UPDATE_DOMAINS_FIELD_NDIM_   5
-#include "cfms_interfaces.inc"
-
