+++ libcFMS/Makefile.am
@@ -31,7 +31,7 @@ lib_LTLIBRARIES = libcFMS.la
 libcFMS_la_LDFLAGS = -version-info 21:0:0
 
 # Add the convenience libraries to the FMS library.
-libcFMS_la_LIBADD = $(top_builddir)/cfms/libcfms.la
+libcFMS_la_LIBADD = $(top_builddir)/c_fms/lib_c_fms.la
 libcFMS_la_LIBADD += $(top_builddir)/c_constants/lib_c_constants.la
 libcFMS_la_LIBADD += $(top_builddir)/c_horiz_interp/lib_c_horiz_interp.la
 libcFMS_la_LIBADD += $(top_builddir)/c_grid_utils/lib_c_grid_utils.la
