+++ test_cfms/c_fms/test_define_domains.sh
