+++ test_cfms/c_fms/test_getset_domains.sh
