+++ test_cfms/c_fms/test_update_domains.sh
